# Faster join CTE