--- a/server/src/instant/db/datalog.clj
+++ b/server/src/instant/db/datalog.clj
@@ -1485,13 +1485,19 @@
                                    :table (:from query)}))
                       (let [join-sym (get-in pattern-group [:children :join-sym])
                             join-cte [(kw prefix next-idx)
-                                      {:select [[[:distinct :entity-id] (kw prefix next-idx :-entity-id)]]
-                                       :from [:triples (kw prefix (dec next-idx))]
-                                       :where (list* :and
-                                                     [:= :app-id app-id]
-                                                     (join-conds prefix
-                                                                 symbol-map
-                                                                 {:e [:variable join-sym]}))}
+                                      (let [conds (join-conds prefix
+                                                              symbol-map
+                                                              {:e [:variable join-sym]})]
+                                        (if-let [single-field (when (and (= 1 (count conds))
+                                                                         (= [:= :entity-id] (take 2 (first conds))))
+                                                                (last (first conds)))]
+                                          {:select [[[:distinct single-field] (kw prefix next-idx :-entity-id)]]
+                                           :from (kw prefix (dec next-idx))}
+                                          {:select [[[:distinct :entity-id] (kw prefix next-idx :-entity-id)]]
+                                           :from [:triples (kw prefix (dec next-idx))]
+                                           :where (list* :and
+                                                         [:= :app-id app-id]
+                                                         conds)}))
                                       :materialized]
                             child-res (accumulate-nested-match-query (-> next-acc
                                                                          (update :ctes conj join-cte)